alphabet generator

+++ CUDA_helpers.cuh
@@ -39,11 +39,11 @@ struct CUDA_Array
     T* CUDA_data;
     size_t num;
 
-    __host__ __device__ inline T& operator[](uint32_t index) {
+    __host__ __device__ inline T& operator[](size_t index) {
         assert(index < num && "Index is out of range in CUDA_Array");
         return CUDA_data[index];
     }
-    __host__ __device__ inline const T& operator[](uint32_t index) const {
+    __host__ __device__ inline const T& operator[](size_t index) const {
         assert(index < num && "Index is out of range in CUDA_Array");
         return CUDA_data[index];
     }
